requirements.txt
# OpenAI
openai==1.109.1

# Prepro + Detect + Recog
paddleocr==3.3.1
paddlepaddle==3.1.1
numpy==1.26.4
pillow==10.2.0
pymupdf==1.24.2
opencv-python==4.9.0.80
torch==2.4.1
torchvision==0.19.1
transformers==4.44.2

# Finetune
transformers==4.44.2
datasets==4.0.0
accelerate==1.11.0
sentencepiece==0.2.1